# docs/openapi/fibness_api.yaml
openapi: 3.0.0
info:
  title: Fibness API
  version: "1.0.0"
servers:
  - url: 'http://10.4.41.146:3001'
    description: FIB server, UPC VPN required to connect
...
paths:
  /user:
    post:
      summary: Creates a user
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: Oriol
                email:
                  type: string
                  example: oriol@example.com
                password:
                  type: string
                  description: Hashed user password
                  example: '5f4dcc3b5aa765d61d8327deb882cf99'
              required:
                - nombre
                - password
                - email
      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: boolean
                    enum: [true]
                  id:
                    type: integer
                    example: 12345
                required:
                  - created
                  - id
        400:
          description: >-
            Some required parameters are missing or with an incorrect type or a
            user with the given email already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: boolean
                    enum: [false]
                  error:
                    type: string
                    example: null value in column 'email' violates constraint
                required:
                  - created 
                  - error
  /user/fb:
    post:
      summary: Log in a user with Facebook token
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: Oriol
                email:
                  type: string
                  example: oriol@example.com
                password:
                  type: string
                  description: Facebook token
                  example: '5f4dcc3b5aa765d61d8327deb882cf99'
              required:
                - nombre
                - password
                - email
      responses:
        200:
          description: User is correct
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 12345
                required:
                  - id
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 12345
                required:
                  - id
  /user/{id}:
    delete:
      summary: Deletes a user by id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to delete
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Ok
  /user/validate:
    post:
      summary: Validates the given user
      description: Validates a user with the given email and password exists
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: oriol@example.com
                password:
                  type: string
                  description: Hashed user password
                  example: '5f4dcc3b5aa765d61d8327deb882cf99'
              required:
                - email
                - password
      responses:
        200:
          description: Result from the validation
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    example: 1234
                required:
                  - result
  /user/{id}/trainings:
    get:
      summary: Gives the trainings of a user
      description: Returns all the trainings of the user with the given id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the trainigs from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of the ids, names and descriptions of the requested trainings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idelemento:
                      type: integer
                    nombre:
                      type: string
                    descripcion:
                      type: string
                    nlikes:
                      type: integer
                    ncomentarios:
                      type: integer
                example:
                  - idelemento: 28
                    nombre: TrainingTest
                    descripcion: DescriptionTest
                    nlikes: 20
                    ncomentarios: 2
                  - idelemento: 40
                    nombre: TrainingTest2
                    descripcion: DescriptionTest2
                    nlikes: 20
                    ncomentarios: 2
                required:
                - trainingSet
  /user/{id}/diets:
    get:
      summary: Gives the diets of a user
      description: Returns all the diets of the user with the given id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the diets from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of the ids, names and descriptions of the requested diets
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    idelemento:
                      type: integer
                    nombre:
                      type: string
                    descripcion:
                      type: string
                    nlikes:
                      type: integer
                    ncomentarios:
                      type: integer
                example:
                  - idelemento: 23
                    nombre: DietTest
                    descripcion: DescriptionTest
                    nlikes: 20
                    ncomentarios: 2
                  - idelemento: 69
                    nombre: DietTest2
                    descripcion: DescriptionTest2
                    nlikes: 20
                    ncomentarios: 2
                required:
                - dietSet
  /user/{id}/routes:
    get:
      summary: Gives the routes of a user 
      description: Returns all the routes of the user with given id
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user, whose's routes we want to get
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Set of the ids, names and descriptions and others of the requested routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  idelemento:
                    type: integer
                  nombre:
                    type: string
                  descripcion:
                    type: string
                  origen:
                    type: string
                  destino:
                    type: string
                  distancia:
                    type: string
                  nlikes:
                    type: integer
                  ncomentarios:
                    type: integer
                example:
                  - idelemento: 3
                    nombre: routeTest
                    descripcion: routeDescription
                    origen: 42.471615;2.271684
                    destino: 45.471615;5.271684
                    distancia: 60
                    nlikes: 20
                    ncomentarios: 2
                    
                  - idelemento: 4
                    nombre: routeTest2
                    descripcion: routeDescription2
                    origen: 43.471615;3.271684
                    destino: 46.471615;6.271684
                    distancia: 70
                    nlikes: 20
                    ncomentarios: 2
                    
                required:
                      - activitySet
  /user/resetPassword:
    put:
      summary: Replaces the current password of the user.
      description: Changes the password of the given user by the new one given
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: oriol@example.com
                password:
                  type: string
                  description: Hashed user password
                  example: '5f4dcc3b5aa765d61d8327deb882cf99'
              required:
                - email
                - password
      responses:
        200:
          description: Result from the reset (true if it was reseted, false otherwise)
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
                required:
                  - result
  /user/{id}/info:
    get:
      summary: Gets the info of the user
      description: Returns the information of the given user
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the information from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Information about the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  nombre:
                    type: string
                    example: UserTest
                  email:
                    type: string
                    example: user@example.com
                  tipousuario:
                    type: string
                    example: principiante
                  tipoperfil:
                    type: string
                    example: publico
                  genero:
                    type: boolean
                    example: true
                  descripcion:
                    type: string
                    example: Esta es mi descripcion
                  fechadenacimiento:
                    type: string
                    example: 1978-01-30
                  fecharegistro:
                    type: string
                    example: Thu Apr 29 2020
                  pais:
                    type: integer
                    example: 54
                  rutaimagen:
                    type: string
                    example: /home/marc/Escritorio/nodeTest
                  nseguidores:
                    type: integer
                    example: 37
                  nseguidos:
                    type: integer
                    example: 55
                  npost:
                    type: integer
                    example: 100
                required:
                  - nombre
                  - email
                  - tipousuario
                  - tipoperfil
                  - genero
                  - descripcion
                  - fechadenacimiento
                  - fecharegistro
                  - pais
                  - rutaimagen
                  - nseguidores
                  - nseguidos
                  - npost
    put:
      summary: Update user info
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the information from
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  example: MiNuevoNombre
                genero:
                  type: boolean
                  example: false
                descripcion:
                  type: string
                  example: Mi nueva descripcion
                fechadenacimiento:
                  type: string
                  example: 1997-12-30
                pais:
                  type: integer
                  example: 56
              required:
                - nombre
                - genero
                - descripcion
                - fechadenacimiento
                - pais
      responses:
        200:
          description: User is correct
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 12345
                required:
                  - id
  /user/{id}/profile:
    get:
      summary: Get the user's profile image
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user which profile image should be updated
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: User's profile image in base64
          content:
            image/*:
              schema:
                format: bas64
    post:
      summary: Set a new profile image
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: User's id
          schema:
            type: integer
            minimum: 1
      requestBody:
        description: Profile image in base64
        required: true
        content:
          image/*:
            schema:
              format: base64
      responses:
        201:
          description: Created
  /user/{id}/settings:
    get:
      summary: Gets the settings of the user
      description: Returns the settings of the given user
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the settings from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Settings of the user
          content:
            application/json:
              schema:
                type: object
                properties:
                  sedad:
                    type: boolean
                    example: false
                  sdistancia:
                    type: boolean
                    example: true
                  sinvitacion:
                    type: boolean
                    example: true
                  sseguidor:
                    type: boolean
                    example: false
                  nmensaje:
                    type: boolean
                    example: false
                required:
                  - info
    put:
      summary: Updates the settings of the user
      description: Changes the settings of the user by the ones given
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the settings from
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sEdad:
                  type: boolean
                  example: false
                sDistancia:
                  type: boolean
                  example: true
                sInvitacion:
                  type: boolean
                  example: true
                sSeguidor:
                  type: boolean
                  example: false
                nMensaje:
                  type: boolean
                  example: false
              required:
                - sEdad
                - sDistancia
                - sInvitacion
                - sSeguidor
                - nMensaje
      responses:
        200:
          description: OK
  /user/shortInfo/{currentID}:
    get:
      summary: Gets the brief info of all users
      description: Returns username and id of all users except for the one that is given and the ones that are blocked by the user
        - user
      tags:
        - user
      parameters:
        - name: currentID
          in: path
          required: true
          description: The id of the user to NOT get the info from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Information of the users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 5
                    nombre:
                      type: string
                      example: user
                    bloqueado:
                      type: boolean
                      example: true
                example:
                  - id: 7
                    nombre: user
                    bloqueado: false
                  - id: 8
                    nombre: user2
                    bloqueado: true
  /user/{id}/followers:
    get:
      summary: Gets the followers of a user
      description: Returns the followers of a given user
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the followers from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Information of the users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 5
                    nombre:
                      type: string
                      example: user
                example:
                  - id: 7
                    nombre: user
                  - id: 8
                    nombre: user2
  /user/{id}/followed:
    get:
      summary: Gets the people followed by a user
      description: Returns all the people that the given user follows
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the user to get the people who follows from
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Information of the users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 5
                    nombre:
                      type: string
                      example: user
                example:
                  - id: 7
                    nombre: user
                  - id: 8
                    nombre: user2
  '/user/follow':
    post:
      summary: Creates a following
      description: The user with idFollower will start following the user with idFollowed
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idFollower:
                  type: integer
                  example: 1
                  description: id of the user that will start following
                idFollowed:
                  type: integer
                  example: 2
                  description: id of the user that will be followed by the first one
              required:
                - idFollower
                - idFollowed
      responses:
        '201':
          description: Attempted to create a following
          content:
            application/json:
              schema:
                type: object
                properties:
                  isBlocked:
                    type: boolean
                    example: true
                    description: true if one of the users has the other blocked (in which case the following
                                  is not created). False if there is no blocking (the following is created)

                required:
                  - isBlocked
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/user/follow/{idFollower}/{idFollowed}':
    delete:
      summary: Deletes a following
      description: The user with idFollower stops following the user with idFollowed 
      tags:
        - user
      parameters:
        - name: idFollower
          in: path
          required: true
          description: The id of the user that started to follow
          schema:
            type: integer
            minimum: 1
        - name: idFollowed
          in: path
          required: true
          description: The id of the user that was followed
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Following successfully eliminated
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/user/block':
    post:
      summary: Blocks a user
      description: The user with idBlocker will block the user with idBlocked
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idBlocker:
                  type: integer
                  example: 1
                  description: id of the user that will be blocking
                idBlocked:
                  type: integer
                  example: 2
                  description: id of the user that will be blocked
              required:
                - idBlocker
                - idBlocked
      responses:
        '201':
          description: Attempted to create a blocking
          content:
            application/json:
              schema:
                type: object
                properties:
                  blockerFollowedBlocked:
                    type: boolean
                    example: true
                    description: true if idBlocker followed idBlocked (the following is eliminated). False
                                  otherwise
                  blockedFollowedBlocker:
                    type: boolean
                    example: false
                    description: true if idBlocked followed idBlocker (the following is eliminated). False
                                  otherwise
                required:
                  - blockerFollowedBlocked
                  - blockedFollowedBlocker
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/user/block/{idBlocker}/{idBlocked}':
    delete:
      summary: Unblocks a user
      description: The user with idBlocker will stop blocking the user with idBlocked 
      tags:
        - user
      parameters:
        - name: idBlocker
          in: path
          required: true
          description: The id of the user that had blocked
          schema:
            type: integer
            minimum: 1
        - name: idBlocked
          in: path
          required: true
          description: The id of the user that has beedn blocked
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Blocking removed
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
...
  '/user/like':
    post:
      summary: Likes an element or a comment
      description: The user will indicate that it likes a comment or an element (training, route or diet)
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idUser:
                  type: integer
                  example: 1
                  description: id of the user that will like something
                idElement:
                  type: integer
                  example: 2
                  description: id of the element OR COMMENT that wil be liked
                type:
                  type: string
                  example: "comment"
                  description: indicates if the idElement corresponds to an element (training, diet or route)
                                or to a comment (type = "element" or "comment" respectively)
              required:
                - idUser
                - idElement
                - type
      responses:
        '201':
          description: Like created successfully
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
  '/user/like/{idUser}/{idElement}/{type}':
    delete:
      summary: Unlikes a comment or element
      description: The user with indicate that it no longer likes an element or comment 
      tags:
        - user
      parameters:
        - name: idUser
          in: path
          required: true
          description: The id of the user will stop licking
          schema:
            type: integer
            minimum: 1
        - name: idElement
          in: path
          required: true
          description: The id of the elemento that no longer will be licked by the user
          schema:
            type: integer
            minimum: 1
        - name: type
          in: path
          required: true
          description: indicates whether the idElement references an element or comment
          schema:
            type: string
      responses:
        '200':
          description: Like removed
        '400':
          description: >-
            Some required parameters are missing or with an incorrect type
...